refactor(tui): extract session-info equality and abort-notice seams for the LOC ratchet
